requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
conda==4.8.3
conda-package-handling==1.7.0
cryptography==2.9.2
dj-static==0.0.6
Django==3.0.5
django-globals==0.3.2
djangorestframework==3.12.2
djongo==1.3.3
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.9
isort==5.6.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongoengine==0.20.0
Pillow==8.0.1
postgres==3.0.0
promise==2.3
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycosat==0.6.3
pycparser==2.20
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymongo==3.11.0
pyOpenSSL==19.1.0
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.55
requests==2.23.0
ruamel-yaml==0.15.87
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.2.4
static3==0.7.0
toml==0.10.1
tqdm==4.46.0
Unidecode==1.1.1
urllib3==1.25.8
wrapt==1.12.1
